Rename collector to importer
Also rename all the importing functions
